Module aws_sdk_appintegrations::operation

Structs

CreateEventIntegration

Creates an EventIntegration, given a specified name, description, and a reference to an

DeleteEventIntegration

Deletes the specified existing event integration. If the event integration is associated

GetEventIntegration

Return information about the event integration.

ListEventIntegrationAssociations

Returns a paginated list of event integration associations in the account.

ListEventIntegrations

Returns a paginated list of event integrations in the account.

ListTagsForResource

Lists the tags for the specified resource.

TagResource

Adds the specified tags to the specified resource.

UntagResource

Removes the specified tags from the specified resource.

UpdateEventIntegration

Updates the description of an event integration.